transfer serviceTime to logfilter

+++ dapeng-core/src/main/java/com/github/dapeng/core/InvocationContext.java
@@ -212,7 +212,7 @@ public interface InvocationContext {
          *
          * @return
          */
-        int serviceTime();
+        long serviceTime();
 
         /**
          * 负载均衡策略
